app/Services/ChartServices/requirements.txt
mysql-connector-python
pandas
numpy
matplotlib
mplfinance
stockstats
pandas_datareader